+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.2.1
 
 - fixed disable derail on game start
+- fixed crash on some train cars
 
 # 1.2.0